Change type of argument for initForMark() method to unsigned int
git-svn-id: https://svn.code.sf.net/p/plexil/code/branches/plexil-3-new-exp@3733 d2be67b0-814c-0410-949f-d215b6b29573

+++ src/intfc/QueueEntry.cc
@@ -84,7 +84,7 @@ namespace PLEXIL
     type = Q_ADD_LIBRARY;
   }
 
-  void QueueEntry::initForMark(uintptr_t seq)
+  void QueueEntry::initForMark(unsigned int seq)
   {
     sequence = seq;
     type = Q_MARK;
